--- chapter_01/hello-angular/app/app.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule  } from '@angular/platform-browser';
-
-import { HelloAngularComponent } from './hello.component';
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [HelloAngularComponent],
-  bootstrap: [HelloAngularComponent],
-})
-export class AppModule { }
