+++ chapter_01/hello-angular/app/main.ts
@@ -1,6 +1,27 @@
+import { Component, NgModule } from '@angular/core';
+import { BrowserModule  } from '@angular/platform-browser';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app.module';
+// Our first "Hello world" component
+@Component({
+  selector: 'hello-angular',
+  template: '<h1 class="text-center"> {{greeting}} </h1>'
+})
+export class HelloAngularComponent {
+  greeting: string;
+  constructor() {
+    this.greeting = 'Hello Angular 2!';
+  }
+}
 
+// Main module, bootstrapping HelloAngularComponent as root component
+@NgModule({
+  imports: [BrowserModule],
+  declarations: [HelloAngularComponent],
+  bootstrap: [HelloAngularComponent],
+})
+export class AppModule { }
+
+// Application bootstrap (specific for browser environments)
 const platform = platformBrowserDynamic();
 platform.bootstrapModule(AppModule);
