+++ chapter_01/pomodoro-timer/app/main.ts
@@ -1,6 +1,70 @@
+import { Component, NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app.module';
+// Pomodoro timer component
+@Component({
+  selector: 'pomodoro-timer',
+  template: `
+    <div class="text-center">
+      <img src="./assets/img/pomodoro.png" alt="Pomodoro">
+      <h1> {{ minutes }}:{{ seconds | number: '2.0' }} </h1>
+      <p>
+        <button (click)="togglePause()"
+          class="btn btn-danger">
+          {{ buttonLabel }}
+        </button>
+      </p>
+    </div>
+`
+})
+export class PomodoroTimerComponent {
+  minutes: number;
+  seconds: number;
+  isPaused: boolean;
+  buttonLabel: string;
 
+  constructor() {
+    this.resetPomodoro();
+    setInterval(() => this.tick(), 1000);
+  }
+
+  resetPomodoro(): void {
+    this.isPaused = true;
+    this.minutes = 24;
+    this.seconds = 59;
+    this.buttonLabel = 'Start';
+  }
+
+  private tick(): void {
+    if (!this.isPaused) {
+      this.buttonLabel = 'Pause';
+
+      if (--this.seconds < 0) {
+        this.seconds = 59;
+        if (--this.minutes < 0) {
+          this.resetPomodoro();
+        }
+      }
+    }
+  }
+
+  togglePause(): void {
+    this.isPaused = !this.isPaused;
+    if (this.minutes < 24 || this.seconds < 59) {
+      this.buttonLabel = this.isPaused ? 'Resume' : 'Pause';
+    }
+  }
+}
+
+// Main module, bootstrapping PomodoroTimerComponent as root component
+@NgModule({
+  imports: [BrowserModule],
+  declarations: [PomodoroTimerComponent],
+  bootstrap: [PomodoroTimerComponent],
+})
+export class AppModule { }
+
+// Application bootstrap (specific for browser environments)
 const platform = platformBrowserDynamic();
 platform.bootstrapModule(AppModule);
